GAPI (fluid): Erode/Dilate optimization: CPU dispatcher
